refactor dropdown and update test
