--- src/components/Dropdown/index.tsx
@@ -1,18 +1,0 @@
-import { Dispatch, SetStateAction } from "react";
-
-import { DROPDOWN_LINKS } from "./DropdownConstants";
-
-import { DropdownPresentation } from "./Presentation";
-
-export default function Dropdown({
-  setIsDropdownVisible,
-}: {
-  setIsDropdownVisible: Dispatch<SetStateAction<boolean>>;
-}) {
-  return (
-    <DropdownPresentation
-      dropdownLinks={DROPDOWN_LINKS}
-      setIsDropdownVisible={setIsDropdownVisible}
-    />
-  );
-}
